mlmodel-backend/requirements.txt
flask
flask-cors
numpy
pillow
opencv-python
tensorflow
werkzeug
scikit-learn
pandas
joblib
sentence-transformers
torch
faiss-cpu
